requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
numpy==1.24.3
boto3==1.34.0
pyarrow==14.0.2
s3fs==2023.12.0
pyyaml==6.0.1 